+++ spec/x-terminal-spec.js
@@ -45,4 +45,69 @@ describe('x-terminal', () => {
 			expect(moveDown).toHaveBeenCalledWith(1)
 		})
 	})
+
+	describe('unfocus()', () => {
+		it('focuses atom-workspace', async () => {
+			jasmine.attachToDOM(atom.views.getView(atom.workspace))
+			await xTerminalInstance.activate()
+			const model = await xTerminalInstance.openInCenterOrDock(atom.workspace)
+			await model.initializedPromise
+			await model.element.createTerminal()
+
+			expect(model.element).toHaveFocus()
+			xTerminalInstance.unfocus()
+			expect(model.element).not.toHaveFocus()
+		})
+	})
+
+	describe('runCommands()', () => {
+		let activeTerminal, newTerminal, commands
+		beforeEach(() => {
+			activeTerminal = {
+				element: {
+					initializedPromise: Promise.resolve(),
+				},
+				runCommand: jasmine.createSpy('activeTerminal.runCommand'),
+			}
+			newTerminal = {
+				element: {
+					initializedPromise: Promise.resolve(),
+				},
+				runCommand: jasmine.createSpy('newTerminal.runCommand'),
+			}
+			commands = [
+				'command 1',
+				'command 2',
+			]
+			spyOn(xTerminalInstance, 'getActiveTerminal').and.returnValue(activeTerminal)
+			spyOn(xTerminalInstance, 'open').and.returnValue(newTerminal)
+		})
+
+		it('runs commands in new terminal', async () => {
+			await xTerminalInstance.runCommands(commands)
+
+			expect(xTerminalInstance.getActiveTerminal).not.toHaveBeenCalled()
+			expect(newTerminal.runCommand).toHaveBeenCalledWith('command 1')
+			expect(newTerminal.runCommand).toHaveBeenCalledWith('command 2')
+		})
+
+		it('runs commands in active terminal', async () => {
+			atom.config.set('x-terminal.terminalSettings.runInActive', true)
+			await xTerminalInstance.runCommands(commands)
+
+			expect(xTerminalInstance.open).not.toHaveBeenCalled()
+			expect(activeTerminal.runCommand).toHaveBeenCalledWith('command 1')
+			expect(activeTerminal.runCommand).toHaveBeenCalledWith('command 2')
+		})
+
+		it('runs commands in new terminal if none active', async () => {
+			xTerminalInstance.getActiveTerminal.and.returnValue()
+			atom.config.set('x-terminal.terminalSettings.runInActive', true)
+			await xTerminalInstance.runCommands(commands)
+
+			expect(xTerminalInstance.getActiveTerminal).toHaveBeenCalled()
+			expect(newTerminal.runCommand).toHaveBeenCalledWith('command 1')
+			expect(newTerminal.runCommand).toHaveBeenCalledWith('command 2')
+		})
+	})
 })